fixed @mattaniBunny's terrible mistake that almost cost me my life, a life-sized pure gold statue of A.H, and all of Apple's shares

+++ IDPS/ArpTable.cpp
@@ -80,5 +80,5 @@ void ArpTable::writeTableToFile()
 
     // Write each IP&MAC pair as a new line
     for (const auto& [ip, mac] : m_table)
-        tableFile << Helper::ipToStr(ip) << ',' << mac.macToString() << '\n';
+        tableFile << Helper::ipToStr(static_cast<ULONG>(ip)) << ',' << mac.macToString() << '\n';
 }
